refactor(Notification): 외부 알림 전송 로직 패키지 구조 변경 및 슬랙 알림 전송 로직과의 결합도 완화
