+++ src/main/java/page/clab/api/domain/auth/accountLockInfo/application/service/MemberBanService.java
@@ -11,8 +11,8 @@ import page.clab.api.domain.auth.accountLockInfo.domain.AccountLockInfo;
 import page.clab.api.domain.memberManagement.member.application.dto.shared.MemberBasicInfoDto;
 import page.clab.api.external.auth.redisToken.application.port.ExternalManageRedisTokenUseCase;
 import page.clab.api.external.memberManagement.member.application.port.ExternalRetrieveMemberUseCase;
-import page.clab.api.global.common.slack.application.SlackService;
-import page.clab.api.global.common.slack.domain.SecurityAlertType;
+import page.clab.api.global.common.notificationSetting.adapter.out.slack.SlackService;
+import page.clab.api.global.common.notificationSetting.domain.SecurityAlertType;
 
 @Service
 @RequiredArgsConstructor
@@ -30,7 +30,7 @@ public class MemberBanService implements BanMemberUseCase {
      * <p>해당 멤버의 계정 잠금 정보를 조회하고, 없으면 새로 생성합니다.
      * Redis에 저장된 해당 멤버의 인증 토큰을 삭제하며, Slack에 밴 알림을 전송합니다.</p>
      *
-     * @param request 현재 요청 객체
+     * @param request  현재 요청 객체
      * @param memberId 차단할 멤버의 ID
      * @return 저장된 계정 잠금 정보의 ID
      */
@@ -58,6 +58,7 @@ public class MemberBanService implements BanMemberUseCase {
 
     private void sendSlackBanNotification(HttpServletRequest request, String memberId) {
         String memberName = externalRetrieveMemberUseCase.getMemberBasicInfoById(memberId).getMemberName();
-        slackService.sendSecurityAlertNotification(request, SecurityAlertType.MEMBER_BANNED, "ID: " + memberId + ", Name: " + memberName);
+        slackService.sendSecurityAlertNotification(request, SecurityAlertType.MEMBER_BANNED,
+                "ID: " + memberId + ", Name: " + memberName);
     }
 }
