+++ src/main/java/page/clab/api/domain/auth/accountLockInfo/application/service/MemberUnbanService.java
@@ -10,8 +10,8 @@ import page.clab.api.domain.auth.accountLockInfo.application.port.out.RetrieveAc
 import page.clab.api.domain.auth.accountLockInfo.domain.AccountLockInfo;
 import page.clab.api.domain.memberManagement.member.application.dto.shared.MemberBasicInfoDto;
 import page.clab.api.external.memberManagement.member.application.port.ExternalRetrieveMemberUseCase;
-import page.clab.api.global.common.slack.application.SlackService;
-import page.clab.api.global.common.slack.domain.SecurityAlertType;
+import page.clab.api.global.common.notificationSetting.adapter.out.slack.SlackService;
+import page.clab.api.global.common.notificationSetting.domain.SecurityAlertType;
 
 @Service
 @RequiredArgsConstructor
@@ -28,7 +28,7 @@ public class MemberUnbanService implements UnbanMemberUseCase {
      * <p>해당 멤버의 계정 잠금 정보를 조회하고 해제합니다.
      * 해제된 정보는 저장되며, Slack에 해제 알림이 전송됩니다.</p>
      *
-     * @param request 현재 요청 객체
+     * @param request  현재 요청 객체
      * @param memberId 해제할 멤버의 ID
      * @return 업데이트된 계정 잠금 정보의 ID
      */
@@ -55,6 +55,7 @@ public class MemberUnbanService implements UnbanMemberUseCase {
 
     private void sendSlackUnbanNotification(HttpServletRequest request, String memberId) {
         String memberName = externalRetrieveMemberUseCase.getMemberBasicInfoById(memberId).getMemberName();
-        slackService.sendSecurityAlertNotification(request, SecurityAlertType.MEMBER_UNBANNED, "ID: " + memberId + ", Name: " + memberName);
+        slackService.sendSecurityAlertNotification(request, SecurityAlertType.MEMBER_UNBANNED,
+                "ID: " + memberId + ", Name: " + memberName);
     }
 }
