+++ src/main/java/page/clab/api/domain/auth/blacklistIp/application/service/BlacklistIpRemoveService.java
@@ -8,8 +8,8 @@ import page.clab.api.domain.auth.blacklistIp.application.port.in.RemoveBlacklist
 import page.clab.api.domain.auth.blacklistIp.application.port.out.RemoveBlacklistIpPort;
 import page.clab.api.domain.auth.blacklistIp.application.port.out.RetrieveBlacklistIpPort;
 import page.clab.api.domain.auth.blacklistIp.domain.BlacklistIp;
-import page.clab.api.global.common.slack.application.SlackService;
-import page.clab.api.global.common.slack.domain.SecurityAlertType;
+import page.clab.api.global.common.notificationSetting.adapter.out.slack.SlackService;
+import page.clab.api.global.common.notificationSetting.domain.SecurityAlertType;
 
 @Service
 @RequiredArgsConstructor
@@ -25,7 +25,7 @@ public class BlacklistIpRemoveService implements RemoveBlacklistIpUseCase {
      * <p>블랙리스트에 등록된 IP 주소 정보를 조회하고 해당 정보를 삭제합니다.
      * 삭제가 완료되면 Slack을 통해 보안 알림이 전송됩니다.</p>
      *
-     * @param request 현재 요청 객체
+     * @param request   현재 요청 객체
      * @param ipAddress 제거할 블랙리스트 IP 주소
      * @return 삭제된 블랙리스트 IP 주소
      */
@@ -34,7 +34,8 @@ public class BlacklistIpRemoveService implements RemoveBlacklistIpUseCase {
     public String removeBlacklistIp(HttpServletRequest request, String ipAddress) {
         BlacklistIp blacklistIp = retrieveBlacklistIpPort.getByIpAddress(ipAddress);
         removeBlacklistIpPort.delete(blacklistIp);
-        slackService.sendSecurityAlertNotification(request, SecurityAlertType.BLACKLISTED_IP_REMOVED, "Deleted IP: " + ipAddress);
+        slackService.sendSecurityAlertNotification(request, SecurityAlertType.BLACKLISTED_IP_REMOVED,
+                "Deleted IP: " + ipAddress);
         return blacklistIp.getIpAddress();
     }
 }
