+++ src/main/java/page/clab/api/domain/auth/redisIpAccessMonitor/application/service/AbnormalAccessIpsClearService.java
@@ -1,6 +1,7 @@
 package page.clab.api.domain.auth.redisIpAccessMonitor.application.service;
 
 import jakarta.servlet.http.HttpServletRequest;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -8,10 +9,8 @@ import page.clab.api.domain.auth.redisIpAccessMonitor.application.port.in.ClearA
 import page.clab.api.domain.auth.redisIpAccessMonitor.application.port.out.ClearIpAccessMonitorPort;
 import page.clab.api.domain.auth.redisIpAccessMonitor.application.port.out.RetrieveIpAccessMonitorPort;
 import page.clab.api.domain.auth.redisIpAccessMonitor.domain.RedisIpAccessMonitor;
-import page.clab.api.global.common.slack.application.SlackService;
-import page.clab.api.global.common.slack.domain.SecurityAlertType;
-
-import java.util.List;
+import page.clab.api.global.common.notificationSetting.adapter.out.slack.SlackService;
+import page.clab.api.global.common.notificationSetting.domain.SecurityAlertType;
 
 @Service
 @RequiredArgsConstructor
@@ -26,7 +25,8 @@ public class AbnormalAccessIpsClearService implements ClearAbnormalAccessIpsUseC
     public List<RedisIpAccessMonitor> clearAbnormalAccessIps(HttpServletRequest request) {
         List<RedisIpAccessMonitor> ipAccessMonitors = retrieveIpAccessMonitorPort.findAll();
         clearIpAccessMonitorPort.deleteAll();
-        slackService.sendSecurityAlertNotification(request, SecurityAlertType.ABNORMAL_ACCESS_IP_DELETED, "Deleted IP: ALL");
+        slackService.sendSecurityAlertNotification(request, SecurityAlertType.ABNORMAL_ACCESS_IP_DELETED,
+                "Deleted IP: ALL");
         return ipAccessMonitors;
     }
 }
