+++ src/main/java/page/clab/api/domain/community/board/application/service/BoardRegisterService.java
@@ -1,5 +1,6 @@
 package page.clab.api.domain.community.board.application.service;
 
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -8,16 +9,14 @@ import page.clab.api.domain.community.board.application.dto.request.BoardRequest
 import page.clab.api.domain.community.board.application.port.in.RegisterBoardUseCase;
 import page.clab.api.domain.community.board.application.port.out.RegisterBoardPort;
 import page.clab.api.domain.community.board.domain.Board;
-import page.clab.api.global.common.slack.domain.SlackBoardInfo;
 import page.clab.api.domain.memberManagement.member.application.dto.shared.MemberDetailedInfoDto;
 import page.clab.api.external.memberManagement.member.application.port.ExternalRetrieveMemberUseCase;
 import page.clab.api.external.memberManagement.notification.application.port.ExternalSendNotificationUseCase;
 import page.clab.api.global.common.file.application.UploadedFileService;
 import page.clab.api.global.common.file.domain.UploadedFile;
-import page.clab.api.global.common.slack.application.SlackService;
+import page.clab.api.global.common.notificationSetting.adapter.out.slack.SlackService;
+import page.clab.api.global.common.slack.domain.SlackBoardInfo;
 import page.clab.api.global.exception.PermissionDeniedException;
-
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -48,7 +47,8 @@ public class BoardRegisterService implements RegisterBoardUseCase {
         Board board = mapper.fromDto(requestDto, currentMemberInfo.getMemberId(), uploadedFiles);
         board.validateAccessPermissionForCreation(currentMemberInfo);
         if (board.shouldNotifyForNewBoard(currentMemberInfo)) {
-            externalSendNotificationUseCase.sendNotificationToMember(currentMemberInfo.getMemberId(), "[" + board.getTitle() + "] 새로운 공지사항이 등록되었습니다.");
+            externalSendNotificationUseCase.sendNotificationToMember(currentMemberInfo.getMemberId(),
+                    "[" + board.getTitle() + "] 새로운 공지사항이 등록되었습니다.");
         }
         SlackBoardInfo boardInfo = SlackBoardInfo.create(board, currentMemberInfo);
         slackService.sendNewBoardNotification(boardInfo);
