+++ src/main/java/page/clab/api/external/auth/accountLockInfo/port/ExternalAccountLockManagementService.java
@@ -13,8 +13,8 @@ import page.clab.api.domain.auth.login.application.exception.MemberLockedExcepti
 import page.clab.api.domain.memberManagement.member.application.dto.shared.MemberDetailedInfoDto;
 import page.clab.api.external.auth.accountLockInfo.application.ExternalManageAccountLockUseCase;
 import page.clab.api.external.memberManagement.member.application.port.ExternalRetrieveMemberUseCase;
-import page.clab.api.global.common.slack.application.SlackService;
-import page.clab.api.global.common.slack.domain.SecurityAlertType;
+import page.clab.api.global.common.notificationSetting.adapter.out.slack.SlackService;
+import page.clab.api.global.common.notificationSetting.domain.SecurityAlertType;
 
 @Service
 @RequiredArgsConstructor
@@ -39,7 +39,7 @@ public class ExternalAccountLockManagementService implements ExternalManageAccou
      *
      * @param memberId 잠금 해제하려는 멤버의 ID
      * @throws MemberLockedException 계정이 현재 잠겨 있을 경우 예외 발생
-     * @throws LoginFailedException 멤버가 존재하지 않을 경우 예외 발생
+     * @throws LoginFailedException  멤버가 존재하지 않을 경우 예외 발생
      */
     @Transactional
     @Override
@@ -55,17 +55,17 @@ public class ExternalAccountLockManagementService implements ExternalManageAccou
      * 로그인 실패를 처리하고 계정 잠금을 관리합니다.
      *
      * <p>로그인 실패 시 멤버의 존재 여부와 계정 잠금 상태를 확인합니다.
-     * 로그인 실패 횟수를 증가시키며, 설정된 최대 실패 횟수에 도달하면 계정을 잠그고
-     * Slack에 보안 알림을 전송합니다.</p>
+     * 로그인 실패 횟수를 증가시키며, 설정된 최대 실패 횟수에 도달하면 계정을 잠그고 Slack에 보안 알림을 전송합니다.</p>
      *
-     * @param request 현재 HTTP 요청 객체
+     * @param request  현재 HTTP 요청 객체
      * @param memberId 로그인 실패를 기록할 멤버의 ID
      * @throws MemberLockedException 계정이 현재 잠겨 있을 경우 예외 발생
-     * @throws LoginFailedException 멤버가 존재하지 않을 경우 예외 발생
+     * @throws LoginFailedException  멤버가 존재하지 않을 경우 예외 발생
      */
     @Transactional
     @Override
-    public void handleLoginFailure(HttpServletRequest request, String memberId) throws MemberLockedException, LoginFailedException {
+    public void handleLoginFailure(HttpServletRequest request, String memberId)
+            throws MemberLockedException, LoginFailedException {
         ensureMemberExists(memberId);
         AccountLockInfo accountLockInfo = ensureAccountLockInfo(memberId);
         validateAccountLockStatus(accountLockInfo);
@@ -99,7 +99,8 @@ public class ExternalAccountLockManagementService implements ExternalManageAccou
         String memberName = memberInfo.getMemberName();
         if (memberInfo.isAdminRole()) {
             request.setAttribute("member", memberId + " " + memberName);
-            slackService.sendSecurityAlertNotification(request, SecurityAlertType.REPEATED_LOGIN_FAILURES, "로그인 실패 횟수 초과로 계정이 잠겼습니다.");
+            slackService.sendSecurityAlertNotification(request, SecurityAlertType.REPEATED_LOGIN_FAILURES,
+                    "로그인 실패 횟수 초과로 계정이 잠겼습니다.");
         }
     }
 }
