+++ src/main/java/page/clab/api/global/auth/filter/CustomBasicAuthenticationFilter.java
@@ -4,6 +4,8 @@ import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.util.Base64;
 import lombok.extern.slf4j.Slf4j;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.security.authentication.AuthenticationManager;
@@ -15,21 +17,17 @@ import org.springframework.security.web.authentication.www.BasicAuthenticationFi
 import page.clab.api.external.auth.blacklistIp.application.port.ExternalRetrieveBlacklistIpUseCase;
 import page.clab.api.external.auth.redisIpAccessMonitor.application.port.ExternalCheckIpBlockedUseCase;
 import page.clab.api.global.auth.util.IpWhitelistValidator;
-import page.clab.api.global.common.slack.application.SlackService;
-import page.clab.api.global.common.slack.domain.SecurityAlertType;
+import page.clab.api.global.common.notificationSetting.adapter.out.slack.SlackService;
+import page.clab.api.global.common.notificationSetting.domain.SecurityAlertType;
 import page.clab.api.global.util.HttpReqResUtil;
 import page.clab.api.global.util.ResponseUtil;
 import page.clab.api.global.util.WhitelistPathMatcher;
-
-import java.io.IOException;
-import java.util.Base64;
 
 /**
  * {@code CustomBasicAuthenticationFilter}는 기본 인증 필터를 확장하여 추가적인 보안 기능을 제공합니다.
  *
  * <p>IP 주소 기반 접근 제한, 화이트리스트 경로 검증, 사용자 인증 정보를 바탕으로
- * Slack 보안 알림을 전송하는 기능을 포함합니다. 또한 Swagger 또는 Actuator에 대한
- * 접근이 성공하거나 실패할 경우 이를 Slack에 알립니다.</p>
+ * Slack 보안 알림을 전송하는 기능을 포함합니다. 또한 Swagger 또는 Actuator에 대한 접근이 성공하거나 실패할 경우 이를 Slack에 알립니다.</p>
  *
  * <p>이 필터는 다음과 같은 추가 검증을 수행합니다:</p>
  * <ul>
@@ -65,6 +63,13 @@ public class CustomBasicAuthenticationFilter extends BasicAuthenticationFilter {
         this.slackService = slackService;
     }
 
+    @NotNull
+    private static String[] decodeCredentials(String authorizationHeader) {
+        String base64Credentials = authorizationHeader.substring("Basic ".length());
+        String credentials = new String(Base64.getDecoder().decode(base64Credentials));
+        return credentials.split(":", 2);
+    }
+
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain chain)
             throws IOException, ServletException {
@@ -82,7 +87,8 @@ public class CustomBasicAuthenticationFilter extends BasicAuthenticationFilter {
         super.doFilterInternal(request, response, chain);
     }
 
-    private boolean authenticateUserCredentials(HttpServletRequest request, HttpServletResponse response) throws IOException {
+    private boolean authenticateUserCredentials(HttpServletRequest request, HttpServletResponse response)
+            throws IOException {
         String authorizationHeader = request.getHeader("Authorization");
         if (authorizationHeader == null || !authorizationHeader.startsWith("Basic ")) {
             response.setHeader("WWW-Authenticate", "Basic realm=\"Please enter your username and password\"");
@@ -122,28 +128,25 @@ public class CustomBasicAuthenticationFilter extends BasicAuthenticationFilter {
         return true;
     }
 
-    @NotNull
-    private static String[] decodeCredentials(String authorizationHeader) {
-        String base64Credentials = authorizationHeader.substring("Basic ".length());
-        String credentials = new String(Base64.getDecoder().decode(base64Credentials));
-        return credentials.split(":", 2);
-    }
-
     private void sendAuthenticationSuccessAlertSlackMessage(HttpServletRequest request) {
         String path = request.getRequestURI();
         if (WhitelistPathMatcher.isSwaggerIndexEndpoint(path)) {
-            slackService.sendSecurityAlertNotification(request, SecurityAlertType.API_DOCS_ACCESS,"API 문서에 대한 접근이 허가되었습니다.");
+            slackService.sendSecurityAlertNotification(request, SecurityAlertType.API_DOCS_ACCESS,
+                    "API 문서에 대한 접근이 허가되었습니다.");
         } else if (WhitelistPathMatcher.isActuatorRequest(path)) {
-            slackService.sendSecurityAlertNotification(request, SecurityAlertType.ACTUATOR_ACCESS,"Actuator에 대한 접근이 허가되었습니다.");
+            slackService.sendSecurityAlertNotification(request, SecurityAlertType.ACTUATOR_ACCESS,
+                    "Actuator에 대한 접근이 허가되었습니다.");
         }
     }
 
     private void sendAuthenticationFailureAlertSlackMessage(HttpServletRequest request) {
         String path = request.getRequestURI();
         if (WhitelistPathMatcher.isSwaggerIndexEndpoint(path)) {
-            slackService.sendSecurityAlertNotification(request, SecurityAlertType.API_DOCS_ACCESS,"API 문서에 대한 접근이 거부되었습니다.");
+            slackService.sendSecurityAlertNotification(request, SecurityAlertType.API_DOCS_ACCESS,
+                    "API 문서에 대한 접근이 거부되었습니다.");
         } else if (WhitelistPathMatcher.isActuatorRequest(path)) {
-            slackService.sendSecurityAlertNotification(request, SecurityAlertType.ACTUATOR_ACCESS,"Actuator에 대한 접근이 거부되었습니다.");
+            slackService.sendSecurityAlertNotification(request, SecurityAlertType.ACTUATOR_ACCESS,
+                    "Actuator에 대한 접근이 거부되었습니다.");
         }
     }
 }
