+++ src/main/java/page/clab/api/global/auth/filter/InvalidEndpointAccessFilter.java
@@ -6,6 +6,7 @@ import jakarta.servlet.ServletRequest;
 import jakarta.servlet.ServletResponse;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.Authentication;
@@ -14,17 +15,14 @@ import org.springframework.web.filter.GenericFilterBean;
 import page.clab.api.domain.auth.blacklistIp.domain.BlacklistIp;
 import page.clab.api.external.auth.blacklistIp.application.port.ExternalRegisterBlacklistIpUseCase;
 import page.clab.api.external.auth.blacklistIp.application.port.ExternalRetrieveBlacklistIpUseCase;
-import page.clab.api.global.common.slack.application.SlackService;
-import page.clab.api.global.common.slack.domain.SecurityAlertType;
+import page.clab.api.global.common.notificationSetting.adapter.out.slack.SlackService;
+import page.clab.api.global.common.notificationSetting.domain.SecurityAlertType;
 import page.clab.api.global.util.HttpReqResUtil;
 import page.clab.api.global.util.ResponseUtil;
 import page.clab.api.global.util.SecurityPatternChecker;
 
-import java.io.IOException;
-
 /**
- * {@code InvalidEndpointAccessFilter}는 서버 내부 파일 및 디렉토리에 대한 비정상적인 접근을 차단하고
- * 보안 경고를 전송하는 필터입니다.
+ * {@code InvalidEndpointAccessFilter}는 서버 내부 파일 및 디렉토리에 대한 비정상적인 접근을 차단하고 보안 경고를 전송하는 필터입니다.
  *
  * <p>특정 패턴을 통해 비정상적인 접근 시도를 탐지하며, 비정상적인 경로로 접근을 시도한 IP를
  * 블랙리스트에 등록하고, Slack을 통해 보안 경고 메시지를 전송합니다.</p>
@@ -50,7 +48,8 @@ public class InvalidEndpointAccessFilter extends GenericFilterBean {
     private final ExternalRetrieveBlacklistIpUseCase externalRetrieveBlacklistIpUseCase;
 
     @Override
-    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws IOException, ServletException {
+    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain)
+            throws IOException, ServletException {
         HttpServletRequest httpRequest = (HttpServletRequest) request;
         String path = httpRequest.getRequestURI();
         boolean isUploadedFileAccess = path.startsWith(fileURL);
@@ -75,7 +74,8 @@ public class InvalidEndpointAccessFilter extends GenericFilterBean {
 
     private void logSuspiciousAccess(HttpServletRequest request, String clientIpAddress) {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        String id = (authentication == null || authentication.getName() == null) ? "anonymous" : authentication.getName();
+        String id =
+                (authentication == null || authentication.getName() == null) ? "anonymous" : authentication.getName();
         String requestUrl = request.getRequestURI();
         String httpMethod = request.getMethod();
         int statusCode = HttpServletResponse.SC_FORBIDDEN;
@@ -98,6 +98,7 @@ public class InvalidEndpointAccessFilter extends GenericFilterBean {
         String blacklistAddedMessage = "Added IP: " + clientIpAddress;
 
         slackService.sendSecurityAlertNotification(request, SecurityAlertType.ABNORMAL_ACCESS, abnormalAccessMessage);
-        slackService.sendSecurityAlertNotification(request, SecurityAlertType.BLACKLISTED_IP_ADDED, blacklistAddedMessage);
+        slackService.sendSecurityAlertNotification(request, SecurityAlertType.BLACKLISTED_IP_ADDED,
+                blacklistAddedMessage);
     }
 }
