+++ src/main/java/page/clab/api/global/common/notificationSetting/application/exception/AlertTypeNotFoundException.java
@@ -1,0 +1,8 @@
+package page.clab.api.global.common.notificationSetting.application.exception;
+
+public class AlertTypeNotFoundException extends RuntimeException {
+
+    public AlertTypeNotFoundException(String alertTypeName) {
+        super("Unknown alert type: " + alertTypeName);
+    }
+}
