+++ src/main/java/page/clab/api/global/common/notificationSetting/domain/AlertType.java
@@ -1,0 +1,8 @@
+package page.clab.api.global.common.notificationSetting.domain;
+
+public interface AlertType {
+
+    String getTitle();
+
+    String getDefaultMessage();
+}
