--- src/main/java/page/clab/api/global/common/slack/dto/mapper/SlackDtoMapper.java
@@ -1,16 +1,0 @@
-package page.clab.api.global.common.slack.dto.mapper;
-
-import org.springframework.stereotype.Component;
-import page.clab.api.global.common.slack.domain.NotificationSetting;
-import page.clab.api.global.common.slack.dto.response.NotificationSettingResponseDto;
-
-@Component
-public class SlackDtoMapper {
-
-    public NotificationSettingResponseDto toDto(NotificationSetting setting) {
-        return NotificationSettingResponseDto.builder()
-                .alertType(setting.getAlertType().getTitle())
-                .enabled(setting.isEnabled())
-                .build();
-    }
-}
