--- src/main/java/page/clab/api/global/common/slack/exception/AlertTypeNotFoundException.java
@@ -1,8 +1,0 @@
-package page.clab.api.global.common.slack.exception;
-
-public class AlertTypeNotFoundException extends RuntimeException {
-
-    public AlertTypeNotFoundException(String alertTypeName) {
-        super("Unknown alert type: " + alertTypeName);
-    }
-}
